requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
httpx
markdown